// src/app/features/reportBuilder/components/queryBuilder/types.ts
export interface QueryBuilderOptions {
  addRuleToNewGroups: boolean;
  autoSelectField: boolean;
  autoSelectOperator: boolean;
  showCombinatorsBetweenRules: boolean;
  debugMode: boolean;
  disabled: boolean;
  enableDragAndDrop: boolean;
  independentCombinators: boolean;
  listsAsArrays: boolean;
  parseNumbers: boolean;
  resetOnFieldChange: boolean;
  resetOnOperatorChange: boolean;
  showNotToggle: boolean;
  showCloneButtons: boolean;
  showLockButtons: boolean;
  showShiftActions: boolean;
  suppressStandardClassnames: boolean;
  useValidation: boolean;
}